refactor(recovery): extract job helpers to keep recovery.ts under 400 lines
Moves countJobsForRun, requeueMissingProbes, and queueSummarizeJobsForRecovery
into a new recovery-jobs.ts module. No behaviour change — satisfies CI
file-size gate.
